(SERVER-3050) Allow retrieving facts from any terminus
Previously, the v4 catalog endpoint would only fetch facts (as opposed
to requiring them in the request) if the PuppetDB facts terminus was
configured. This caused issues for people wanting to use extensions of
the PDB terminus, like satellite, because the endpoint would error
claiming that PDB wasn't configured.

This commit removes the restriction on the fact retrieval logic. Now, if
facts are not supplied in the request body, puppetserver will attempt to
fetch them from whatever facts terminus is configured, whether that is
yaml/json (default), PuppetDB, or some custom terminus. So long as the
terminus implements `find`, this will succeed.