(SERVER-3050) Bypass the fact cache in the v4 catalog endpoint
When we stopped explicitly querying the PuppetDB terminus for facts, we
inadventently started using the yaml cache terminus instead of PuppetDB,
even when PuppetDB was configured. However, trusted facts are not stored
in the yaml cache, and the primary use case for this endpoint requires
stored trusted facts to be retrieved from the system. So this commit
updates the fact retrieval to bypass the cache terminus for facts, going
instead directly to the primary terminus.

When PuppetDB is configured, this means we will query PDB for facts,
rather than the default yaml cache, and therefore retrieve its stored
trusted facts.