(maint) Update ca-certificates
Recently a major root certificate expired and had to be replaced. This
caused postgres to fail to install, on platforms where we install PDB.
This commit adds a step to our presuite that fetches new CA certificates,
to ensure that next time a cert is rotated, we will fetch it automatically.