(maint) Stop testing with PuppetDB on EL6
The recent root certificate expiration has made it difficult to install
postgres on EL6. RedHat did not issue an update of the `ca-certificates`
package for thie EOL platform, so avoiding the expired cert is much more
complicated. Since we will soon be dropping support for EL6, we decided
to instead stop testing the PDB integration on this platform. This
commit removes it from the list of PDB master platforms. If a good
workaround is found, we can put this back.