(SERVER-3080) Error intelligibly if CRL is missing auth key id
Previously, if a CRL without an authority key identifier extension was
submitted to the CRL update endpoint, we would throw an unintelligible
Clojure schema error instead of telling the user what went wrong. This
commit adds a check that ensures the supplied CRLs have auth key ids,
and throws a meaningful exception which will return a 400 from the
endpoint if they do not.