(SERVER-3079) Run dropsonde weekly
This commit adds a new scheduled job to the analytics service, to submit
module metrics to dujour via dropsonde. This submission is managed by
the same service as the existing analytics submission, but is configured
separately and runs weekly instead of daily.

Submission is disabled by default. When enabled, metrics will be
collected and submitted when the server starts, and weekly thereafter.
A failed dropsonde invocation will log an error but otherwise cause no
issues.