(SERVER-3068) Take up new clj-parent with new http-client
The last update to puppetlabs/http-client brought in a version that did
not forward auth related headers when being redirected. This was
problematic and two things occurred in the newest version of http-client
(1.2.3):

  1. The definition of "auth" header was expanded to include cookie and
     proxy-auth* headers.
  2. The filtering of auth headers will not occur if the redirection
     points to a uri which matches the original's scheme, host, and port.

With these changes all the internal Puppet projects should be able to
take up this version. Consequently, it was merged to clj-parent and now
we may remove our http-client pin and update clj-parent.

Also the version of comidi will be bumped as a side effect of taking up
this version of clj-parent.