(SERVER-3079) Use clj-shell-utils for shelling out to Dropsonde
Elsewhere in the server, we use our own shell execution library to
invoke other processes. This gives us control over the interface and
lets us ensure that we're not doing bad things like forking the JVM when
we shell out. This commit updates the Dropsonde invocation to use the
same utility.