(SERVER-3079) Install scanf gem explicitly
Ruby 2.7 stopped bundling `scanf` in its stdlib, but JRuby still does.
So when we install `dropsonde` with JRuby, it doesn't bother to install
the transitive `scanf` dep (used by Puppet). Then when we run
`dropsonde` under MRI, the `scanf` dep is missing. This commit adds
`scanf` to the list of gems to be explicitly installed.