(SERVER-3114) Do not put SANs on the CA signing cert
This commit removes the SAN we added to the self-signed CA signing cert
as part of SERVER-2338. The common name of this certificate is not a
valid DNS name, which means some tools, like LibreSSL will error when
validating it, since we had labled it as a DNS name in the extension.
The name itself is fine in this context since it's not used to validate
connections, just sign other certificates. But for the same reason, this
cert does not need alternate identities in the first place.