pick first certificate
If there are other certificates in $SSLDIR/certs/, then this script
fails on the last line.
With this change, the script just picks the first certificate it finds,
that is not named ca.pem. The expected behavior when having multiple
certificates is unknown to me, but like that it a least doesn't fail.