(maint) Ensure needed container directories are owned by puppet
The container is created with the puppet user owning these directories
and if a user causes the UID of the puppet user to change than they will
not be readable.

This forces those directories to be owned by the puppet user regardless
of UID on start up.

h/t to jay7x who originally proposed this.