(SERVER-2662) Use existing cacert's subject key id when signing
With the release of jvm-ssl-utils 3.4.0, the authority key identifer
can be copied from an existing certificate and used when signing
CSRs, instead of always computing a type 1 key from the CA's public
key. This commit changes puppetserver to use the CA's certificate
during signing so that puppetserver signs CSRs with the existing CA's
subject key id as the authority key id for agent requests. This will
allow puppetserver to use externally created CA certificates that
use a type 2 key.