(maint) Add custom merge driver script
This commit adds a script that can be used as a merge driver for project.clj
and for the beaker options file. It will resolve common conflicts around version
bumps, but any unfamiliar conflicts will need manual intervention. This script
was copied from pe-puppet-server-extensions and modified slightly to handle the
beaker options file. Previously, puppetserver mergeup conflicts were handled by
a custom merge driver created as part of the `puppetserver-mergeup.sh` script in
ci-job-configs, but this did not handle conflicts in project.clj. This strategy
will allow us to resolve fewer conflicts by hand and will allow us to simplify
the mergeup script in jenkins.