(PE-33991) remove use of slurp to reduce memory footprint
This removes the use of slurp in the clojure code to instead
present the input stream to ruby to process. While ruby will
still read the string when the body is requested, this change
will prevent the clj layer from having a referenced string during
the lifetime of the request. This should help reduce memory overhead.