(maint) Conditionally open Java IO only for Java 17
In a previous commit we began passing flags to the JVM that allows
JRuby to access protected members of the Java IO subsystem, as is
currently required by JRuby 9.3.8. However, this flag does not exist
in (at least) Java 8 and so we must conditionally include these flags
based on Java versions.

The conditional code is based on the existing conditional JVM flags
passed in FIPS mode.