(SERVER-3167) Update to JRuby 9.4 work around Timeout issue
JRuby's terminate function will premeturely stop shutting down threads
if it comes across an adopted thread. This will cause threads not
properly shutdown by the time a JRuby instance terminates to be leaked.

This may happen with code in the wild. Within Puppet Server it only
happens with the Timeout module which is poorly behaved in this regard.
This update should allow us to begin testing JRuby 9.4.x and Ruby 3.x
compatability in Puppet Enterprise now, but we should not Puppet Server
with this change. We must either take up the fix to this thread leakage,
merged in https://github.com/jruby/jruby/pull/7633 slated to be released
in JRuby 9.4.1.1 or 9.4.2.0 depending on the number and severity of
additional bugs reported, or revert this update.

Note: the JRuby team estimates 2 - 8 weeks between when the above PR was
merged and a new release containing it, which would be sometime between
late February to early April 2023.