(PE-35325) Update Puppet Server to include the requester's info when logging CA actions
Added an optional dependency for the ActivityReportingService, as well as new functionality for attempting to extract CA signee data from requests to sign and revoke certificates. With preference for RBAC user details, but will fallback to the name linked to the certificate used to sign, and defaulting to the “CA” signee if no other information is available. When a certificate event occurs, log messages will now include information on who signed the certificate. Additionally, if the activity service is available, the information will be submitted there as well.