(maint) Remove unneeded Timeout module workaround
In JRuby 9.4.0 & 9.4.1 there was an issue when shutting down threads,
if JRuby encountered an adopted thread it would stop its thread shutdown
procedure and any abandoned threads would be leaked.

This was particularly problematic because the Timeout module does not
attempt to shutdown its watcher thread and Puppet Server adopts a Jetty
thread as part of processing incoming requests, so we would consistently
leak Timeout threads when flushing a JRuby.

To work around this issue while we tested our JRuby upgrade we manually
killed the Timeout watcher thread. This was not a vialable solution to
ship with, but allowed us to see how much work the move to a Ruby 3
compatible would be.

With JRuby 9.4.2 the issue is resolved and we can remove the workaround.