(SERVER-3246) Add lein-eastwood linting
There are three main exceptions to eastwood's defaults as well as a few
classes of changes that eastwood necessitated:

Disabled checks:
 - Reflection: There's a ton of reflection happening and in general it
   would be great to eliminate this with type hints. There may even be
   performance improvements from doing so. However, we should do this as
   separate work. Orchestrator, which also uses eastwood, disables
   reflection checking.
 - Second argument to "is" isn't a string. We often call a pretty
   printer or display the incorrect subject for easier debugging rather
   than a string explaining the failure. eastwood recommends wrapping
   such a function call in "str", which is asinine.
 - Unused meta on macro. Some i18n macros expand to include unnecessary
   metadata on them. Not our concern

Classes of changes:
 - A surprising number of tests didn't use "is" (and just had a
   boolean expression like (= :foo "foo")).
 - Some def/defns had docstrings in the wrong location.
 - Several methods in libraries and Java itself have been deprecated and
   are updated with modern versions.

The biggest change comes from removing the deprecated
"puppetlabs.dujour.version-check/check-for-updates!" which would execute
the check in a background thread. This meant when the check fails in the
exception and messaging would be lost. The new function now executes in
the calling thread so we now wrap it in a try block, catch Exception and
log an error. This is a better behavior but causes much more noise in
our tests.