(PE-35637) add explicit read/write locking to serial file use
In preparation for future changes coming to the CA endpoints, this
adds explicit read/write locking to the management of the serial file.

This also adds the ability for the serial lock acquisition to timeout
preventing potential locking issues. The default timeout is 5 seconds.

The locks are added to the ca-settings, and schemas updated as appropriate.