(PE-35638) ensure get-crl-last-modified is protected by locks
This adds read locking to the `get-crl-last-modified` routine to prevent
the value from being read while the file is written.  It also updates
`handle-get-certificate-revocation-list` to get a read lock to prevent
a window where the crl modification time is read, but the file was written
later.

Additionally a small optimization was added to only get the modification time
for the crl if the if-modified-since header is provided.