(PE-35640) ensure inventory append operations are crash-safe
Prior to this change, it was possible that a partial inventory
write, or changes to the inventory file could occur to cause corruption.

This uses the atomic-write behavior to ensure that only whole inventory
files will be present in the location where they are expected. The
new behavior uses the kitchen-sink atomic-write with a streaming function
to allow the current content to be streamed to a new file, the new entry
appended, and once that is done, moved to replace the old file in
an atomic way.

A test was added to demonstrate the correct handling of large inventory
file content.