(SERVER-3288) ensure reader is closed with use of read-infra-nodes
In a code inspection it was noted that the reader used in in the
`read-infra-nodes` funciton in the certificate-authority would leak
an open reader after the sequence was evaluated. This alters the
function signature to pass in the reader, and allow the caller
to manage the lifetime of the reader as appropriate.