(PE-35638) consolidate crl-locks into one, simplify lock passing
This consolidates the crl-locks into one instead of two, as that
is how the crl is conceptually managed. The routines int he
certificate-authority space that were passed the locks were consolidated
and treated as effectively private functions, assuming that locking is
done outside of them.

Tests were updated as needed.