(PE-35709) add in renew-certificate function
Given an existing certificate, which is assumed to be from the
same issuer (caller to validate), generate a new signed cert
with the same extensions, and a new expiration period determined
by the auto-renewal-cert-ttl time.

Note that the issuer and subject identifier extensions need to be
regenerated for signing to occur properly.

Tests are added to demonstrate the behavior of the function.