(PE-35705) extract certificate from renewal request
This commit updates the handle-cert-renewal function to extract
certificates present in the request. Certificates can be included
either as a ssl-client-cert property in the request, or as an
x-client-cert field in the request header if the allow-header-cert-info
config option is true.

Imports for certificate-authority-core and certificate-authority-int-test
were alphabetized as well.