(PE-36859) ensure revocation endpoints work without cert present
This updates the certificate-status and clean endpoints to introspect the
inventory list for potential matches to the requested cert revocation in
addition to the cert on disk.

The lock for the inventory file is now a read/write lock due to the
need to read the inventory file as well as write to it.

Tests were added to demonstrate the behaviors. The tests fail prior
to the fixes being applied.