(PE-36952) ensure all infrastructure serials are written
This adds a new behavior such that when a renewal or signing is done
and the host in question is in the infrastructure list, the serial number
for that signed cert is immediately written to the infra-serial file.

This has the benefit of ensuring that all certs will be correctly revoked in
the infra-crl when it is updated on revocation.