(PE-37347) add bulk signing function
This adds a new function that given a set of csr names, the CA settings
and a function to report activity will attempt to sign all the csrs
that are valid.  The function returns a map that contains the names of
the csrs that were signed, and those that weren't signed.

The function validates alt-names, authorization-extensions, and signatures.

Locking is done across the group to prevent lock contention.  In local
testing, it took ~5 seconds to process ~500 records, of which 100 were valid.