(maint) disable dropsonde during acceptance tests
It was discovered that the dropsonde service was running during
acceptance tests, and due to the nature of the tests (many reloads
in rapid succession, and lack of outward connectivity), the ruby
processes could cause excessive memory use. This disables the service
during acceptance tests, as we don't need to test it specifically there.