(maint) add license to project.clj
As part of a recent change in clojars, it is necessary for the license
to be part of the exported pom file. Puppetserver has been under the
apache-2 license for ~12 years, so this isn't anything new. This adds
the declaration to the project.clj so that leiningen will correctly add
the license to the exported pom file.