(maint) small optimization for revocation path
In the case that there are no serials to revoke, it is unnecessary
to introspect the infra-nodes to see if they intersect with the
serials to be revoked. This adds a small optimization to avoid
the computation and file reading associated with the infra-nodes
in the case that the serial number set is empty.

It also creates a temporary cache of the ca-public-key and ca-private-key
within the function to avoid reading and extracting it twice potentially.