(issue) only issue GET requests for concrete files
The inventory is managed in several ways in the CA, including an inventory
file, which records certificates, expiration periods and serial numbers.
In the case where we want to get the actual certificate status, we don't want
to use the inventory file. The liberator `exists` method was checking in both
the filesystem and the inventory file for the certificate in question for
a `GET` request.  This alters that behavior to only check the file system under
those circumstances.

A test is added to demonstrate the behavior.

Resolves https://github.com/puppetlabs/puppetserver/issues/2851