(PE-38948) Respect rich_data in compileCatalog
Previously, when using the compileCatalog method of
Puppet::Server::Master compilation would fail to serialize rich data.
This occurred because during serialization we introspect the
Puppet.context for the rich_data value (the setting itself defaults to
true & and strict now defaults to error, throwing an error when we
encounter an object that requires rich_data to serialize).

Normal catalog compilation works because the value of rich_data in the
Puppet.context is overridden by the caller of the serialization code for
indirected routes, using whatever is set in the environment and if unset
in the environment then falling back to the actually rich_data setting.

In the compileCatalog method however we do not use the indirected routes
for serialization, instead calling `to_data_hash` on the catalog and
serializing that ourselves (we explicitly only support JSON via that
endpoint). This means the value in the Puppet.context is the default
value which at the present is `false` regardless of what the user has
configured (or our own settings defaults).

This commit attempts to fix this by overriding the Puppet.context's
rich_data value with the Puppet setting for rich_data. Which now
defaults to true and will honor whatever the user puts in their
puppet.conf. This however does not support rich_data per environment. I
believe the optimal point to override rich_data is at the overall
compileCatalog level but we do not have the "current" environment at
that point, neccessitating a much larger rewrite of the compileCatalog
function to work. The fact that you can even set rich_data in the
environment.conf at this time is an undocumented feature, which leads me
to believe a larger rewrite of the compileCatalog function is not
a good investment.