(PE-38408) Remove expensive Regexes from puppet profiler Java impl
We have triggered a nasty performance regression in the JVM[1] with some
of our Regex usage in the MetricsPuppetProfiler (particularly the
PuppetDB Patterns). Regardless of the performance regression, our Regex
usage is unnecessary and unneedlessly expensive (I believe it was
taking up ~3% of samples from JFR in an unrelated support escalation).

This patch moves from iterating over all metrics and doing regex matches
to pull out only those that match the desired pattern when metrics are
requested, to caching references to desired metrics in maps as they are
recorded. The keys of these maps should be the same as the match group
from the previous patterns.