(maint) Fix code_scripts acceptance test
There is some change for Redhat 9's SSH set up that are causing this
acceptance test to now fail. I thought originally it was a peculiarity
of git shipped on Redhat 9 and an issue with the safe.directory setting.
After wasting half a day on it and finally root causing the true issue,
just moving to use a file protocol is the simplest fix. This test is
looking at the code_scripts extension points, the way the git repo is
configured is not important to the test.